hw_6_prompting/requirements.txt
pandas ~= 2.2.2
pyyaml ~= 6.0.1
openai ~= 1.30.1
fastapi ~= 0.110.0
python-dotenv ~= 1.0.0
langchain~= 0.2.14
faiss-cpu~= 1.8.0
html2text~= 2020.1.16
uvicorn ~= 0.29
tensorflow~=2.16
tensorflow-text~=2.16
tensorflow-hub~=0.16.1
pydantic~=2.5.2
requests~=2.31.0
langchain_community~=0.2.12